+++ src/index.ts
@@ -92,6 +92,8 @@ export type {
   LibraryItem,
   LibraryList,
   LibraryListInput,
+  OnProgress,
+  Progress,
   RelatedVerifications,
   SelectInput,
   SimilarVerification,